+++ src/sentry/rules/conditions/tagged_event.py
@@ -4,10 +4,7 @@ from collections import OrderedDict
 from django import forms
 
 from sentry import tagstore
-from sentry.api.serializers.models.project import bulk_fetch_project_latest_releases
 from sentry.rules.conditions.base import EventCondition
-from sentry.signals import release_created
-from sentry.utils.cache import cache
 
 
 class MatchType(object):
@@ -35,17 +32,6 @@ MATCH_CHOICES = OrderedDict(
 )
 
 
-def get_project_release_cache_key(project_id):
-    return u"project:{}:latest_release".format(project_id)
-
-
-@release_created.connect(weak=False)
-def clear_project_release_cache(release, **kwargs):
-    release_project_ids = release.projects.values_list("id", flat=True)
-    for proj_id in release_project_ids:
-        cache.delete(get_project_release_cache_key(proj_id))
-
-
 class TaggedEventForm(forms.Form):
     key = forms.CharField(widget=forms.TextInput())
     match = forms.ChoiceField(MATCH_CHOICES.items(), widget=forms.Select())
@@ -54,22 +40,11 @@ class TaggedEventForm(forms.Form):
     def clean(self):
         super(TaggedEventForm, self).clean()
 
-        key = self.cleaned_data.get("key")
         match = self.cleaned_data.get("match")
         value = self.cleaned_data.get("value")
 
         if match not in (MatchType.IS_SET, MatchType.NOT_SET) and not value:
             raise forms.ValidationError("This field is required.")
-
-        if (
-            key == "release"
-            and value == "latest"
-            and match
-            not in (MatchType.EQUAL, MatchType.NOT_EQUAL, MatchType.IS_SET, MatchType.NOT_SET)
-        ):
-            raise forms.ValidationError(
-                "When matching on latest release you must use 'equals' or 'does not equal'"
-            )
 
 
 class TaggedEventCondition(EventCondition):
@@ -81,16 +56,6 @@ class TaggedEventCondition(EventCondition):
         "match": {"type": "choice", "choices": MATCH_CHOICES.items()},
         "value": {"type": "string", "placeholder": "value"},
     }
-
-    def get_latest_release(self, event):
-        cache_key = get_project_release_cache_key(event.group.project_id)
-        latest_release = cache.get(cache_key)
-        if not latest_release:
-            latest_releases = bulk_fetch_project_latest_releases([event.group.project])
-            if latest_releases:
-                cache.set(cache_key, latest_releases[0], 600)
-                return latest_releases[0]
-        return latest_release
 
     def passes(self, event, state, **kwargs):
         key = self.get_option("key")
@@ -127,24 +92,6 @@ class TaggedEventCondition(EventCondition):
             for k, v in event.tags
             if k.lower() == key or tagstore.get_standardized_key(k) == key
         )
-
-        # specific case to handle if the user wants to match 'release equals latest'
-        if key == "release" and value == "latest":
-            latest_release = self.get_latest_release(event)
-            if not latest_release:
-                return False
-
-            if match == MatchType.EQUAL:
-                ret_value = True
-            elif match == MatchType.NOT_EQUAL:
-                ret_value = False
-            else:
-                return False
-
-            for t_value in values:
-                if t_value == latest_release.version:
-                    return ret_value
-            return not ret_value
 
         if match == MatchType.EQUAL:
             for t_value in values:
