+++ tests/sentry/rules/conditions/test_tagged_event.py
@@ -1,11 +1,7 @@
 from __future__ import absolute_import
 
-from datetime import datetime
-
 from sentry.testutils.cases import RuleTestCase
 from sentry.rules.conditions.tagged_event import TaggedEventCondition, MatchType
-from sentry.models import Release
-from sentry.signals import release_created
 
 
 class TaggedEventConditionTest(RuleTestCase):
@@ -109,69 +105,3 @@ class TaggedEventConditionTest(RuleTestCase):
 
         rule = self.get_rule(data={"match": MatchType.NOT_SET, "key": "missing"})
         self.assertPasses(rule, event)
-
-    def test_latest_release(self):
-        event = self.get_event()
-        oldRelease = Release.objects.create(
-            organization_id=self.organization.id,
-            version="1",
-            date_added=datetime(2020, 9, 1, 3, 8, 24, 880386),
-        )
-        oldRelease.add_project(self.project)
-
-        newRelease = Release.objects.create(
-            organization_id=self.organization.id,
-            version="2",
-            date_added=datetime(2020, 9, 2, 3, 8, 24, 880386),
-        )
-        newRelease.add_project(self.project)
-
-        event.data["tags"] = (("release", newRelease.version),)
-        rule = self.get_rule(data={"match": MatchType.EQUAL, "key": "release", "value": "latest"})
-        self.assertPasses(rule, event)
-
-    def test_latest_release_no_match(self):
-        event = self.get_event()
-        oldRelease = Release.objects.create(
-            organization_id=self.organization.id,
-            version="1",
-            date_added=datetime(2020, 9, 1, 3, 8, 24, 880386),
-        )
-        oldRelease.add_project(self.project)
-
-        newRelease = Release.objects.create(
-            organization_id=self.organization.id,
-            version="2",
-            date_added=datetime(2020, 9, 2, 3, 8, 24, 880386),
-        )
-        newRelease.add_project(self.project)
-
-        event.data["tags"] = (("release", oldRelease.version),)
-        rule = self.get_rule(data={"match": MatchType.EQUAL, "key": "release", "value": "latest"})
-        self.assertDoesNotPass(rule, event)
-
-    def test_caching(self):
-        event = self.get_event()
-        oldRelease = Release.objects.create(
-            organization_id=self.organization.id,
-            version="1",
-            date_added=datetime(2020, 9, 1, 3, 8, 24, 880386),
-        )
-        oldRelease.add_project(self.project)
-        event.data["tags"] = (("release", oldRelease.version),)
-        rule = self.get_rule(data={"match": MatchType.EQUAL, "key": "release", "value": "latest"})
-        self.assertPasses(rule, event)
-
-        newRelease = Release.objects.create(
-            organization_id=self.organization.id,
-            version="2",
-            date_added=datetime(2020, 9, 2, 3, 8, 24, 880386),
-        )
-        newRelease.add_project(self.project)
-        # confirm the rule still passes on an old release due to caching
-        self.assertPasses(rule, event)
-
-        # now send the appropriate signal to trigger cache clear
-        release_created.send_robust(release=newRelease, sender=self.__class__)
-        # now that cache is cleared rule should not pass on an old release
-        self.assertDoesNotPass(rule, event)
